Fix TypeError: 'float' object cannot be interpreted as an index

Good catch. I suspect there are more of these floating around. Note that testbest4.py is only provided for convenient testing outside of the test suite (you don't need to submit it to t-square) and the library version's on buffet0x are old enough that float indexing still works.

+++ mc3h1_defeat_learners/testbest4.py
@@ -12,7 +12,7 @@ from gen_data import best4LinReg, best4RT
 def compare_os_rmse(learner1, learner2, X, Y):
 
     # compute how much of the data is training and testing
-    train_rows = math.floor(0.6* X.shape[0])
+    train_rows = int(math.floor(0.6* X.shape[0]))
     test_rows = X.shape[0] - train_rows
 
     # separate out training and testing data
